Naive Overall Population Model Estimation

Function to fit the naive overall population model to the data. This model is going to estimate the subgroup treatment effects with the overall population treatment effect.

Usage

naivepop(resp, trt, data, resptype = c("survival", "binary"), status = NULL)

Arguments

resp

(string)
the response variable name.

trt

(string)
the treatment variable name.

data

(data frame)
the data frame with the variables.

resptype

(string)
the type of data used. Can be "survival" or "binary".

status

(string)
only for "survival" resptype, the status variable name in survival data.

Value

List with fit, model, resptype, data.

Examples

naivepop("futime", "trt", survival::myeloid, "survival", "death")
#> $fit
#> Call:
#> survival::coxph(formula = form_surv, data = data)
#> 
#>         coef exp(coef) se(coef)      z       p
#> trtB -0.3457    0.7077   0.1122 -3.081 0.00206
#> 
#> Likelihood ratio test=9.52  on 1 df, p=0.002029
#> n= 646, number of events= 320 
#> 
#> $model
#> [1] "naive_pop"
#> 
#> $resptype
#> [1] "survival"
#> 
#> $data
#>      id trt sex futime death txtime crtime rltime
#> 1     1   B   f    235     1     NA     44    113
#> 2     2   A   m    286     1    200     NA     NA
#> 3     3   A   f   1983     0     NA     38     NA
#> 4     4   B   f   2137     0    245     25     NA
#> 5     5   B   f    326     1    112     56    200
#> 6     6   B   f   2041     0    102     NA     NA
#> 7     7   A   m     63     1     NA     NA     NA
#> 8     8   A   f    446     1    205     34    382
#> 9     9   B   f   1695     0     NA     28     NA
#> 10   10   A   f   1669     0    106     NA     NA
#> 11   11   A   f     66     0     NA     NA     NA
#> 12   12   A   f   1364     0    187     NA     NA
#> 13   13   A   f     17     0     NA     NA     NA
#> 14   14   A   f    209     1     NA     36    144
#> 15   15   A   m    261     1     99     32    154
#> 16   16   B   f   2380     0    195     72     NA
#> 17   17   A   f    431     1    145     NA     NA
#> 18   18   A   m    372     1    153     31     88
#> 19   19   A   f    838     1     NA     32    477
#> 20   20   B   m   2086     0    185     33     49
#> 21   21   A   f   1480     0     NA     41     NA
#> 22   22   B   f    188     1     NA     36    186
#> 23   23   B   f    963     1    101     33    668
#> 24   24   B   f    318     1    271     33    195
#> 25   25   A   f    373     1    117    102     NA
#> 26   26   A   f   1657     0     NA     84     NA
#> 27   27   A   f    213     1     NA     39    184
#> 28   28   A   f   1421     0    208     33    433
#> 29   29   B   f    337     1    251     57    183
#> 30   30   A   m    565     1     NA     36    452
#> 31   31   A   f    286     1    112     NA     NA
#> 32   32   B   m   1823     0    337     29    220
#> 33   33   A   f     17     0     NA     NA     NA
#> 34   34   A   f    201     1     NA     35     NA
#> 35   35   A   f    321     0     NA    117    299
#> 36   36   A   m    529     1    206     55     NA
#> 37   37   B   m    266     1    137     62     NA
#> 38   38   B   m   1305     1    109     31    811
#> 39   39   B   m   1460     0     80     31     NA
#> 40   40   A   m   1832     1     93     NA   1692
#> 41   41   B   f   1859     0     NA     26     NA
#> 42   42   A   m   1864     0    144     29     NA
#> 43   43   B   m   1880     0     NA     36     NA
#> 44   44   B   f    368     1     NA     58    323
#> 45   45   A   m   1391     0    134    111     NA
#> 46   46   B   m    762     1    594     30    426
#> 47   47   B   f    332     1    154     90    144
#> 48   48   B   m    376     0     85     NA     NA
#> 49   49   A   f   1740     0     NA     32     NA
#> 50   50   A   f   2241     0    186     NA     NA
#> 51   51   B   f    582     1     NA     44    282
#> 52   52   A   f    284     1    134     41     77
#> 53   53   B   f    823     1     NA     27    620
#> 54   54   B   f   2186     0    621     96    492
#> 55   55   B   m   2198     0     87     NA     NA
#> 56   56   A   f   1737     0    118     NA     NA
#> 57   57   A   m   2315     0     NA     NA     NA
#> 58   58   B   m   2006     0    242     70     NA
#> 59   59   A   m    185     1     NA     30     NA
#> 60   60   B   m   2367     0     NA     27     NA
#> 61   61   B   f   1893     0     NA     64     NA
#> 62   62   B   m   1734     0    110     31    884
#> 63   63   A   f   2042     0    103     41    261
#> 64   64   B   m   2162     0    137     75    165
#> 65   65   B   f   1176     0     NA     41    995
#> 66   66   A   f   2148     0     NA     30     NA
#> 67   67   B   m   2247     0     90     39     NA
#> 68   68   A   f   1991     0    152     25     NA
#> 69   69   A   m    180     1     NA     NA    131
#> 70   70   B   f   1956     0     NA     31     NA
#> 71   71   B   f   1976     0    164     72     NA
#> 72   72   B   f     13     1     NA     NA     NA
#> 73   73   A   f   1913     0     NA     32     NA
#> 74   74   A   f    110     1     87     NA     NA
#> 75   75   B   f    116     1     NA     26     NA
#> 76   76   A   m    245     1    105     43    168
#> 77   77   B   m   2194     0    180    145     NA
#> 78   78   B   m    531     1    495     29    347
#> 79   79   B   m   1782     0    140     NA     NA
#> 80   80   A   m    548     1     NA     33    507
#> 81   81   B   f     46     1     NA     NA     NA
#> 82   82   A   f   2052     0     NA     29     NA
#> 83   83   B   m    411     1     NA     NA    328
#> 84   84   A   f   1034     1    406     33    320
#> 85   85   A   f    293     1     NA     78    231
#> 86   86   B   m   2250     0     NA     40     NA
#> 87   87   B   f    235     1    108     NA     NA
#> 88   88   A   f    267     1    220     28     NA
#> 89   89   B   m   2136     0    488     64    355
#> 90   90   A   f   1465     0    180     36     NA
#> 91   91   A   m    439     1     NA     52    322
#> 92   92   A   m    295     1    275     28     53
#> 93   93   A   m    286     1    156     21    271
#> 94   94   B   f   1251     0    171     29     NA
#> 95   95   B   m   1360     0     93     35     NA
#> 96   96   A   m   1783     0    244     34     NA
#> 97   97   A   f   1866     0    174     91     NA
#> 98   98   B   m    892     1    371     31    255
#> 99   99   B   f   1525     0     77     48    163
#> 100 100   B   f    461     1     NA     NA     NA
#> 101 101   B   f    322     1     NA     31    204
#> 102 102   B   m   1323     0     92     36     NA
#> 103 103   A   m    651     1    586     50    527
#> 104 104   B   f    241     0     NA     36     NA
#> 105 105   A   f   2300     0    163     NA     NA
#> 106 106   B   m   2192     0    149     50   1905
#> 107 107   B   f    191     1    100     33    170
#> 108 108   A   f     18     0     NA     NA     NA
#> 109 109   B   f   1621     1    664     57    558
#> 110 110   B   m     50     1     NA     NA     NA
#> 111 111   B   m     15     0     NA     NA     NA
#> 112 112   B   f     27     1     NA     NA     NA
#> 113 113   A   f    574     1    101     NA     NA
#> 114 114   B   f    917     1     NA     37    696
#> 115 115   B   m    609     1    369     34    266
#> 116 116   B   m   1774     0     70     NA     NA
#> 117 117   A   m    206     1    131     NA     NA
#> 118 118   A   m    297     1     NA     55     90
#> 119 119   B   f     77     1     53     NA     NA
#> 120 120   B   m     52     1     NA     NA     NA
#> 121 121   B   f   1429     0    103     40     NA
#> 122 122   B   f    278     1    113     52    142
#> 123 123   B   m   1393     0    185     26     NA
#> 124 124   B   m    232     1     NA     73    165
#> 125 125   B   m    106     1     NA     NA     NA
#> 126 126   B   m    397     1     80     NA     NA
#> 127 127   A   m    140     1    100     NA     26
#> 128 128   A   f   2038     0    142     35     NA
#> 129 129   B   f   1911     0     NA     94     NA
#> 130 130   A   f     13     1     NA     NA     NA
#> 131 131   B   m   1990     0     NA     35     NA
#> 132 132   A   m    424     1     NA     35    209
#> 133 133   A   f   2063     0     NA     32     NA
#> 134 134   A   f   1721     0     89     NA     NA
#> 135 135   B   m    955     0     82     34    167
#> 136 136   B   m   1731     0     NA     31     NA
#> 137 137   B   f   1499     0    171     36     NA
#> 138 138   A   f    281     1     NA     NA     NA
#> 139 139   B   m    310     1    105     NA     NA
#> 140 140   B   f   1999     0    155     NA     NA
#> 141 141   A   m    408     1     93     29     NA
#> 142 142   A   f   1361     0    151     40    715
#> 143 143   A   f    337     1    158     32    276
#> 144 144   A   m     15     0     NA     NA     NA
#> 145 145   A   f   1530     0     99     29     NA
#> 146 146   A   m    234     1     79     NA     NA
#> 147 147   B   m   1681     0    120     NA     NA
#> 148 148   A   m    518     1    135     NA    113
#> 149 149   A   f    800     1     NA     74    495
#> 150 150   B   m   1806     0    733     43    634
#> 151 151   A   m   1192     0     NA     44     NA
#> 152 152   A   m   2304     0    128     58     NA
#> 153 153   A   f   1687     0     NA     60     NA
#> 154 154   B   m   1785     0     NA     36     NA
#> 155 155   B   f   1912     0    497     32    417
#> 156 156   A   f   2335     0    175     28    116
#> 157 157   A   f     31     0     NA     21     NA
#> 158 158   A   f    485     1    248     36     NA
#> 159 159   A   m    432     0     77     NA     NA
#> 160 160   A   f   1409     1    444     38    341
#> 161 161   A   f   1464     0     78     70     NA
#> 162 162   A   f     18     1     NA     NA     NA
#> 163 163   A   m    462     1    136     NA    336
#> 164 164   B   m    408     1    230     34    166
#> 165 165   B   f   1423     0     60    186     NA
#> 166 166   B   m   1431     0     NA     NA     NA
#> 167 167   A   f    248     1     NA     38    212
#> 168 168   A   f   1459     0     83     30    193
#> 169 169   A   f   2148     0     62     38    428
#> 170 170   A   f   2188     0     78     NA     NA
#> 171 171   B   m   1725     0     NA     33     NA
#> 172 172   B   f    218     1     NA     50    185
#> 173 173   A   f   1999     0    175    151     NA
#> 174 174   A   m   1254     1     NA     44   1230
#> 175 175   B   m   1420     0     NA     35     NA
#> 176 176   A   f   1521     0     NA     28     NA
#> 177 177   B   m   1921     0    117     29     NA
#> 178 178   A   m   1648     0     NA     54     NA
#> 179 179   A   m   1555     0    182     77     NA
#> 180 180   A   m    614     1    374     57    288
#> 181 181   B   m     50     1     NA     NA     NA
#> 182 182   A   f     86     1     NA     NA     NA
#> 183 183   B   f    588     0     NA     32     NA
#> 184 184   B   m    874     1    703     54    598
#> 185 185   B   m     33     1     NA     NA     NA
#> 186 186   A   f    156     1     NA     32     NA
#> 187 187   B   m    563     1    163     NA     NA
#> 188 188   B   m    231     1    115     NA     NA
#> 189 189   B   m   1020     1    784     56    638
#> 190 190   B   f   1390     0     67     NA     NA
#> 191 191   A   m    572     1     NA     66    402
#> 192 192   B   m   1877     0     NA     33     NA
#> 193 193   B   f   2122     0     NA     33     NA
#> 194 194   A   m    322     1     NA     33    276
#> 195 195   A   f    511     1     NA     98    179
#> 196 196   B   m    288     1     NA     95    241
#> 197 197   B   f    368     1     90     NA     NA
#> 198 198   A   m    480     1    130     NA     NA
#> 199 199   A   f   2224     0    171     NA     NA
#> 200 200   A   f   1708     0    144     NA     NA
#> 201 201   B   f    561     1    139     NA     NA
#> 202 202   A   m   1578     0     NA     38     NA
#> 203 203   B   m    834     1    207     76     93
#> 204 204   B   m   1583     0     NA     48     NA
#> 205 205   B   f      7     0     NA     NA     NA
#> 206 206   B   m    513     1     NA     39    308
#> 207 207   B   f   1426     0    130     35     NA
#> 208 208   B   f   2040     0    988     33    814
#> 209 209   B   m   1796     0     NA     37     NA
#> 210 210   A   f   1627     0    130     35     NA
#> 211 211   A   m    360     1    225     NA     NA
#> 212 212   B   m   1280     0    151     55     NA
#> 213 213   A   f   2371     0     NA     35     NA
#> 214 214   A   m   1479     0     NA    137    257
#> 215 215   A   m    260     1     79     46     NA
#> 216 216   B   f   2219     0    143     NA     NA
#> 217 217   A   m     20     1     NA     NA     NA
#> 218 218   A   m    269     1     98     NA     NA
#> 219 219   B   m    381     1     NA     35    248
#> 220 220   A   m    163     1     NA     34     NA
#> 221 221   B   m   1287     0    137     76     NA
#> 222 222   A   m   1092     0     NA     39    514
#> 223 223   B   f    154     1     NA     28     49
#> 224 224   B   m    771     1    146     NA    338
#> 225 225   B   f   1972     0    656     NA     NA
#> 226 226   B   m   1658     0    150     35     NA
#> 227 227   B   f   2283     1    715     61    681
#> 228 228   A   f   1827     0     90    149     NA
#> 229 229   B   m      4     0     NA     NA     NA
#> 230 230   A   m   2210     0     NA     46     NA
#> 231 231   B   f   1379     1    219     28     NA
#> 232 232   B   m    516     1     NA     73    103
#> 233 233   B   m   1007     1    878     52    654
#> 234 234   A   m   1757     0    126     29     NA
#> 235 235   A   f   2134     0    123     32     NA
#> 236 236   A   m     99     1     NA     NA     NA
#> 237 237   A   f   1369     0     NA     36     NA
#> 238 238   B   m    121     1     NA     NA     NA
#> 239 239   B   m   2361     0    842     47    729
#> 240 240   B   f    423     1    143     NA     NA
#> 241 241   A   f    322     1    254     39    177
#> 242 242   A   f   1423     0    126     NA     NA
#> 243 243   A   m    369     1    121     NA    219
#> 244 244   B   m    603     1    428     38    322
#> 245 245   B   f     25     1     NA     NA     NA
#> 246 246   B   f   1809     0     99     22     NA
#> 247 247   A   m    131     1     NA     63     NA
#> 248 248   B   f    602     1    185     38     NA
#> 249 249   B   f    692     1    361     31    293
#> 250 250   A   f   1444     0    153     36    198
#> 251 251   B   m    601     0    107     35    504
#> 252 252   A   f    305     1     NA     22    242
#> 253 253   A   f    518     1    124    215    369
#> 254 254   B   f   2118     0     NA     43     NA
#> 255 255   B   f    972     1     NA     42    410
#> 256 256   A   f    524     1     NA     49    384
#> 257 257   A   m   1881     0     NA     28     NA
#> 258 258   B   m     74     1     NA     NA     NA
#> 259 259   B   m   2345     0     NA     36     NA
#> 260 260   B   f    403     1    177     33    246
#> 261 261   A   f     14     0     NA     NA     NA
#> 262 262   B   f   2053     0     NA     34     NA
#> 263 263   A   m   2254     0     NA     28     NA
#> 264 264   B   m     76     1     NA     NA     NA
#> 265 265   B   m     22     0     NA     NA     NA
#> 266 266   A   f   2154     0    114     79     NA
#> 267 267   A   m    779     1     NA     62    762
#> 268 268   A   f    167     1     87     62     NA
#> 269 269   A   m     16     0     NA     NA     NA
#> 270 270   A   m    191     1    121     47     75
#> 271 271   A   f    707     1    223     29    124
#> 272 272   B   m    103     1     NA     70     NA
#> 273 273   A   f    283     1     NA     55    220
#> 274 274   B   f    293     1     NA     48    179
#> 275 275   B   m   1442     0    102     NA     NA
#> 276 276   A   f   1350     0    693     28    620
#> 277 277   B   m   1710     0     NA     26     NA
#> 278 278   A   f     19     0     NA     NA     NA
#> 279 279   B   f   1384     0     NA     30     NA
#> 280 280   B   f    263     1     NA     36    240
#> 281 281   A   m   1947     0     NA     54     NA
#> 282 282   B   f    230     1    105    131     NA
#> 283 283   A   f    248     1    214     NA     NA
#> 284 284   A   m    216     1     NA     44     74
#> 285 285   B   f   1717     0     98    214    251
#> 286 286   B   m   1718     0    114     31     NA
#> 287 287   A   f    707     1    451     63    316
#> 288 288   B   f    303     1     NA    176     NA
#> 289 289   A   m     89     1     NA     42     NA
#> 290 290   A   f   1545     0    150     37     NA
#> 291 291   B   m   1283     1     NA     32    440
#> 292 292   B   m    388     1    104     47     NA
#> 293 293   B   f   2023     0    312     37     NA
#> 294 294   A   m   1990     0    150     NA     NA
#> 295 295   A   f    140     1    135     62    123
#> 296 296   B   m    759     1    608     38    535
#> 297 297   A   f   2145     0   1526     NA     NA
#> 298 298   B   f    168     0    143     31     NA
#> 299 299   B   f   1843     0     NA     57     NA
#> 300 300   B   m   2188     0    101     32     NA
#> 301 301   A   m   1960     0     89     NA     NA
#> 302 302   B   f    874     1    403     74    280
#> 303 303   B   f     34     1     NA     NA     NA
#> 304 304   A   m    438     1     72     NA     NA
#> 305 305   A   f    396     1    390     60    284
#> 306 306   A   f    867     1     NA     41    471
#> 307 307   B   m   1885     0     NA     34     NA
#> 308 308   A   f    148     1     NA     NA     NA
#> 309 309   B   f   1522     0    130     28     NA
#> 310 310   B   m   2233     0     NA     67     NA
#> 311 311   B   m   1795     0    274     29     NA
#> 312 312   B   m     13     1     NA     NA     NA
#> 313 313   B   f   1467     0     NA     37     NA
#> 314 314   B   f   1682     0     93     38     NA
#> 315 315   A   f   2063     0     NA     57     NA
#> 316 316   A   f   1302     0    101     73     NA
#> 317 317   A   f     56     1     NA     NA     NA
#> 318 318   A   f   1666     0     97     33     NA
#> 319 319   A   f   1708     0   1292     27   1155
#> 320 320   A   m   1384     0     87     NA     NA
#> 321 321   B   m    644     1     NA     22    612
#> 322 322   A   f    239     1    139     NA     45
#> 323 323   B   f   2095     0    683     36    583
#> 324 324   A   f   2065     0    259     45     NA
#> 325 325   A   f    588     1     NA     41    323
#> 326 326   B   m    432     1    356     37    242
#> 327 327   A   f    308     1     NA     78     NA
#> 328 328   B   f   1779     0    156     NA     NA
#> 329 329   A   m    197     1     NA     60     NA
#> 330 330   B   m   2183     0     NA     67     NA
#> 331 331   A   f   1368     0    101     34     NA
#> 332 332   B   m    193     1     NA     NA     NA
#> 333 333   B   f   2239     0     NA     30     NA
#> 334 334   A   f   1479     0    111     60     NA
#> 335 335   B   m   1414     0     NA     41    648
#> 336 336   B   m    599     1     NA     54    472
#> 337 337   B   f   1834     0    115     58     NA
#> 338 338   A   f   1290     0     NA     88     NA
#> 339 339   B   m   1968     0    704     45    625
#> 340 340   B   m   2123     0    190     84     NA
#> 341 341   A   f    385     1    119     39     NA
#> 342 342   B   m    698     0     NA     38     NA
#> 343 343   B   f   2255     0     NA     74    454
#> 344 344   B   f   1319     0    111     61     NA
#> 345 345   A   m   1465     0    164     28     NA
#> 346 346   A   f   1552     1   1403     29     NA
#> 347 347   A   f   1883     0     89     NA     NA
#> 348 348   A   m    409     1    255     66    188
#> 349 349   A   f    144     1     NA     NA     NA
#> 350 350   A   f    157     1     NA     NA     NA
#> 351 351   A   f   1370     0    284     43    186
#> 352 352   A   f    193     0     86    119     NA
#> 353 353   A   m    128     1     NA     NA     NA
#> 354 354   B   m    671     1    506     28    450
#> 355 355   A   f    670     1    382     56    303
#> 356 356   B   m   1960     0    330     29     NA
#> 357 357   B   f   1320     0     NA     41     NA
#> 358 358   A   f    679     1    238     NA     NA
#> 359 359   B   m   1815     0    168     94     NA
#> 360 360   A   f   2149     0    631     NA     NA
#> 361 361   A   m    748     1     81     28     NA
#> 362 362   B   f   1901     0    741     34    666
#> 363 363   B   f   2229     0     NA     28    190
#> 364 364   A   f    964     1    428    110    375
#> 365 365   B   m   1379     0     NA     NA     NA
#> 366 366   B   f    146     1     NA     54    122
#> 367 367   B   f    179     1     NA     NA     NA
#> 368 368   A   m      6     0     NA     NA     NA
#> 369 369   B   m   1755     0     NA     33     NA
#> 370 370   A   f   2043     0     NA     35     NA
#> 371 371   A   f    258     1    146     33     73
#> 372 372   A   m    460     1    148     NA     NA
#> 373 373   A   f    575     1    202     60    392
#> 374 374   A   f    364     1     77    110     NA
#> 375 375   A   m   2344     0     NA     29     NA
#> 376 376   A   m   1953     0     NA     27     NA
#> 377 377   B   f    386     1     NA     72    276
#> 378 378   A   f   1164     1    161     56    941
#> 379 379   B   f    374     1    214     NA     NA
#> 380 380   B   f   1683     0     NA     41     NA
#> 381 381   B   m   1704     0    185     44     NA
#> 382 382   B   f     86     0     NA     31     NA
#> 383 383   A   m    262     1     NA     NA     NA
#> 384 384   A   m    187     1    119     57     NA
#> 385 385   A   f     65     1     NA     NA     NA
#> 386 386   B   f    355     1    138     45    124
#> 387 387   A   f     36     0     NA     NA     NA
#> 388 388   B   f   1727     0    266     55     NA
#> 389 389   A   f   1855     0     50     NA   1544
#> 390 390   A   m    215     1     NA     59    109
#> 391 391   A   f    319     1    219     NA     NA
#> 392 392   A   m   1861     0     NA     49     NA
#> 393 393   A   f     21     0     NA     NA     NA
#> 394 394   A   m    497     1    191     NA     NA
#> 395 395   A   m   1385     0     NA     40     NA
#> 396 396   B   m      9     1     NA     NA     NA
#> 397 397   B   m    418     1    180     NA     NA
#> 398 398   A   f   2028     0    561     66    491
#> 399 399   A   m    476     1    302     60     NA
#> 400 400   B   m   1744     0     73    217     NA
#> 401 401   B   f   1952     0    166     78     NA
#> 402 402   A   m    344     1    285     57    252
#> 403 403   B   f    464     1     NA     NA     NA
#> 404 404   B   f   1742     0     NA     30     NA
#> 405 405   B   f    355     1     92     47     NA
#> 406 406   B   f   2222     0    149    134     NA
#> 407 407   A   m    226     1     NA     NA     NA
#> 408 408   A   f    356     1     91     NA     NA
#> 409 409   B   f   1405     0     NA     64     NA
#> 410 410   B   f    210     1     NA     64    104
#> 411 411   A   f    907     1    718     30    607
#> 412 412   B   m   2276     0     NA     35     NA
#> 413 413   A   f    104     1     NA     41     87
#> 414 414   B   m    565     1     NA    100    232
#> 415 415   B   f    727     1    530     52    426
#> 416 416   A   m    385     1     78     NA     NA
#> 417 417   B   f     20     1     NA     NA     NA
#> 418 418   B   f    333     1    190     NA     NA
#> 419 419   A   f    578     1     NA     42    299
#> 420 420   B   m   1988     0    780     40     NA
#> 421 421   A   m   1349     0    122     37     NA
#> 422 422   A   f   2007     0     NA     67     NA
#> 423 423   B   f    255     1    212     36    130
#> 424 424   A   m    522     1     NA     NA    418
#> 425 425   A   m    539     1    465     36    328
#> 426 426   A   m     23     1     NA     NA     NA
#> 427 427   B   m   1714     0     NA     NA     NA
#> 428 428   A   m   2056     0     24     24     NA
#> 429 429   A   f    341     1     NA     41    180
#> 430 430   B   m    258     1    131     57    193
#> 431 431   A   m   2237     0    566     84    496
#> 432 432   A   m   1102     1    419     46    236
#> 433 433   B   m   1729     0     98     78    225
#> 434 434   B   f   2318     0     NA     42     NA
#> 435 435   A   f      9     0     NA     NA     NA
#> 436 436   B   m    411     1     NA     NA     NA
#> 437 437   A   m     59     0     NA     38     NA
#> 438 438   B   f    805     1    791     37    467
#> 439 439   A   f   2058     0     NA     54     NA
#> 440 440   B   f    697     1     NA     36     NA
#> 441 441   B   f   2048     0    161     71    691
#> 442 442   B   f   2034     0     NA     73     NA
#> 443 443   A   f    696     1    192    112    153
#> 444 444   B   f    411     1    357     50    288
#> 445 445   A   f   2083     0    236     NA     NA
#> 446 446   A   m    388     1    140     25     45
#> 447 447   B   f    342     1     NA     42    252
#> 448 448   A   f   1544     0     NA     28     NA
#> 449 449   B   f     44     1     NA     NA     NA
#> 450 450   A   f    326     1     94     60    215
#> 451 451   A   m   2200     0     NA     21     NA
#> 452 452   A   f     28     1     NA     NA     NA
#> 453 453   A   m   1360     0    102     37     NA
#> 454 454   B   f    658     1    117     72    458
#> 455 455   B   m    621     1     NA     NA    263
#> 456 456   B   f    307     1    221     44    163
#> 457 457   B   f    248     1    126     NA     NA
#> 458 458   B   f   1652     0     NA     32     NA
#> 459 459   A   f   1623     0    392     45    333
#> 460 460   A   f    540     1     NA    108    358
#> 461 461   B   m    166     1    132     28     NA
#> 462 462   A   f    246     1     NA     NA     NA
#> 463 463   A   f    458     1     84     34    359
#> 464 464   B   m    329     1     NA     49    326
#> 465 465   B   f   2017     0    123     29     NA
#> 466 466   A   m   1129     1     NA     34   1104
#> 467 467   B   f   1720     0    155     44     NA
#> 468 468   B   m   1703     0     NA     49     NA
#> 469 469   A   f    167     1     78     NA     NA
#> 470 470   B   m   1584     0    133    196    441
#> 471 471   A   f      9     0     NA     NA     NA
#> 472 472   B   f   1021     1    476     35    403
#> 473 473   B   m    465     1    247     32     NA
#> 474 474   B   f   1308     0    264     25    190
#> 475 475   B   f   1681     0    145     NA     NA
#> 476 476   A   m   1846     0    321     28     NA
#> 477 477   A   m   1241     1    616     33    567
#> 478 478   B   m   1602     0     NA     35     NA
#> 479 479   A   m   1813     0    235     38     NA
#> 480 480   B   m   1742     0     NA     35     NA
#> 481 481   A   m   1459     0     NA     32     NA
#> 482 482   A   m     95     1     NA     NA     NA
#> 483 483   B   f   1559     0    132     38     NA
#> 484 484   A   f     24     1     NA     NA     NA
#> 485 485   A   f    465     1    363     57    274
#> 486 486   B   f   1333     0   1333     90   1225
#> 487 487   B   f    419     1     93     NA     NA
#> 488 488   B   f   1924     0     NA     35     NA
#> 489 489   A   m     85     1     NA     NA     NA
#> 490 490   A   m   1537     1    134     36     NA
#> 491 491   A   m    246     1    119     NA     NA
#> 492 492   A   m   1774     0    167     22     NA
#> 493 493   B   f   1999     0     NA     29     NA
#> 494 494   B   f   2233     0    559     31    496
#> 495 495   A   f   1367     0     95     NA     NA
#> 496 496   A   f    267     1     NA     38    159
#> 497 497   A   m   1085     1    984     42     NA
#> 498 498   B   f   2052     0    527     40    446
#> 499 499   A   m    742     1    134     NA     NA
#> 500 500   A   f    516     0    128     33     59
#> 501 501   B   m    499     1    409     32    280
#> 502 502   B   f   1907     0    113     47     NA
#> 503 503   A   m    400     1    238     25    170
#> 504 504   A   f   1792     0     NA     62     NA
#> 505 505   B   f   1698     0     NA     NA     NA
#> 506 506   B   m    774     1    220     35    503
#> 507 507   B   f    833     1     NA     NA    238
#> 508 508   A   m   2253     0    128     58     NA
#> 509 509   A   m   1383     0    224     60    177
#> 510 510   B   f   1039     0     NA     82     NA
#> 511 511   A   f    148     1    147     NA     NA
#> 512 512   B   m    447     1     NA     NA    373
#> 513 513   B   m    476     1     69     56    215
#> 514 514   B   m    736     1    582     29    489
#> 515 515   B   m   1485     0     NA     29     NA
#> 516 516   A   m    567     0     92     NA    515
#> 517 517   A   m     27     1     NA     NA     NA
#> 518 518   A   f   1745     0    209     36     NA
#> 519 519   B   m    313     1    225     41     NA
#> 520 520   B   m   1739     0    247     34     NA
#> 521 521   A   f   2043     0     NA     21     NA
#> 522 522   B   m   1760     0    290     61     NA
#> 523 523   B   f     45     1     NA     NA     NA
#> 524 524   B   f    145     1     NA     64     NA
#> 525 525   B   m   1989     0    313     43    238
#> 526 526   B   m    242     1    175     NA     NA
#> 527 527   B   m   1099     0     NA     33     NA
#> 528 528   B   m    506     1    120     28     NA
#> 529 529   A   f   2044     0    582     33    512
#> 530 530   B   m   1497     0    110     21     NA
#> 531 531   A   f    156     1     NA     NA     NA
#> 532 532   A   m    456     1     NA     41    422
#> 533 533   B   m   1695     0     NA     40     NA
#> 534 534   B   f   2299     0     NA     42     NA
#> 535 535   A   m   1281     0     NA     30     NA
#> 536 536   A   f   1990     0    103     33     NA
#> 537 537   B   m     24     1     NA     NA     NA
#> 538 538   B   f    495     1    207     38     NA
#> 539 539   B   f    743     1     63    142    288
#> 540 540   B   f   2419     0    160     38     86
#> 541 541   B   m   1764     0    100     32     NA
#> 542 542   A   f    170     1    135     NA     NA
#> 543 543   B   m   1331     0     NA     34     NA
#> 544 544   B   f    112     1     92     32     NA
#> 545 545   B   f    295     1     NA     NA    120
#> 546 546   A   f    359     1     NA     45     NA
#> 547 547   B   m     32     1     NA     NA     NA
#> 548 548   A   f     27     1     NA     NA     NA
#> 549 549   A   f   1374     0     NA     55     NA
#> 550 550   B   f    120     0     66     NA     NA
#> 551 551   A   f    365     1    253     38    181
#> 552 552   A   f    144     1     72     49     NA
#> 553 553   B   m    168     1    109     NA     NA
#> 554 554   A   f   1901     0     NA     52     NA
#> 555 555   A   m    275     1    113     59    200
#> 556 556   B   m    826     1     NA     29    632
#> 557 557   A   f   1704     0     88     NA     NA
#> 558 558   B   f    432     0    292     22    221
#> 559 559   A   f   1780     0     NA     32     NA
#> 560 560   A   f     49     0     NA     29     NA
#> 561 561   A   f   1313     0     NA     73     NA
#> 562 562   A   f     24     0     NA     NA     NA
#> 563 563   B   m   2126     0     NA     28     NA
#> 564 564   B   f    609     1    204     NA     NA
#> 565 565   A   f    203     1     NA     77    160
#> 566 566   A   m   1245     1    122     77     NA
#> 567 567   A   f    436     0     85     NA     NA
#> 568 568   B   f   1825     0    105     44     NA
#> 569 569   A   f   1915     1     NA     41     NA
#> 570 570   A   f   2150     0    132     NA   1033
#> 571 571   B   f   1710     0     NA    554    928
#> 572 572   B   f   2125     0     NA     NA     NA
#> 573 573   B   m   1594     0     NA     31     NA
#> 574 574   A   m   1414     0     NA     34     NA
#> 575 575   B   m   1455     0    127     40     NA
#> 576 576   A   f   1799     0     NA     26     NA
#> 577 577   A   f   1728     0    122     33     73
#> 578 578   B   f   1688     0    173     NA     NA
#> 579 579   A   m   1113     0     NA     33     NA
#> 580 580   A   m    185     1    109     NA     NA
#> 581 581   B   f     10     0     NA     NA     NA
#> 582 582   A   m    811     1    144     21     NA
#> 583 583   A   f   1480     0     82     NA     NA
#> 584 584   B   f    515     1     83     NA     NA
#> 585 585   A   f    122     1     NA     NA     NA
#> 586 586   A   f   2179     0     NA     67     NA
#> 587 587   B   f    420     1     NA     65    324
#> 588 588   B   f    248     1    115     NA     NA
#> 589 589   A   f     24     1     NA     NA     NA
#> 590 590   B   f    906     0    227     39     NA
#> 591 591   A   m    486     1    341     37    276
#> 592 592   B   f    273     1    130     32     NA
#> 593 593   B   m   1220     1     NA     29    842
#> 594 594   A   m     40     1     NA     NA     NA
#> 595 595   A   f    443     1     NA     41    232
#> 596 596   B   m    383     1    204     NA    132
#> 597 597   B   f    829     1    669     55    497
#> 598 598   A   m    160     0     NA     45     NA
#> 599 599   B   f    275     1    237     28     68
#> 600 600   A   m    400     1     78     NA     NA
#> 601 601   A   f    768     1    433     84    365
#> 602 602   B   m   1347     0    169     38     NA
#> 603 603   B   m   1374     0    400     65    295
#> 604 604   A   f    375     1    213     NA     NA
#> 605 605   B   f   1456     0    213     52     NA
#> 606 606   B   f     27     1     NA     NA     NA
#> 607 607   A   f     93     0     NA     NA     NA
#> 608 608   B   m   1943     0     NA     46     NA
#> 609 609   B   f   2041     0    152     42     NA
#> 610 610   A   m    517     1    107     60    457
#> 611 611   B   f    252     1    134     39     NA
#> 612 612   A   f    293     1    129     31    219
#> 613 613   A   m   1417     0    147     22     NA
#> 614 614   B   m    752     1    319     53     NA
#> 615 615   B   f   1322     0    153     NA     NA
#> 616 616   B   m   2070     0    127     65   1124
#> 617 617   A   m     16     1     NA     NA     NA
#> 618 618   B   m   2191     0    159     39     NA
#> 619 619   A   f    334     1     NA     75    234
#> 620 620   B   f   1370     1    379     35    329
#> 621 621   B   m   2369     0     NA     37     NA
#> 622 622   B   f   1401     0     92     34     NA
#> 623 623   B   m    664     1    642     30    530
#> 624 624   A   m    259     1    117     58     NA
#> 625 625   A   f   2308     0    127     NA     NA
#> 626 626   B   f   2335     0     NA     29     NA
#> 627 627   A   f    120     1     NA     NA     89
#> 628 628   B   f   2055     0    101     77     NA
#> 629 629   B   f   1730     0     NA     34     NA
#> 630 630   B   m   1373     0     89    115     NA
#> 631 631   B   f    563     1    104     NA     NA
#> 632 632   B   m   1480     0    104     76     NA
#> 633 633   B   f    884     1    572     45     NA
#> 634 634   B   m   1386     0    103     50     NA
#> 635 635   B   f   1845     0     NA     93     NA
#> 636 636   B   f   1313     0     82     NA     NA
#> 637 637   A   f    410     1     94     NA     45
#> 638 638   B   m    697     1    434     26    327
#> 639 639   A   m    253     1    155     NA     NA
#> 640 640   A   m     21     1     NA     NA     NA
#> 641 641   A   f     22     0     NA     NA     NA
#> 642 642   B   f    237     1    223     78    191
#> 643 643   A   f   2394     0     94     NA     NA
#> 644 644   A   f     17     0     NA     NA     NA
#> 645 645   A   f    181     1     99     66    144
#> 646 646   B   m     25     0     NA     NA     NA
#> 
#> attr(,"class")
#> [1] "shrinkforest" "naivepop"