Summary Horseshoe Function

Function to obtain the estimated subgroup treatment effects from a horseshoe model and a credible interval for them.

Usage

# S3 method for horseshoe
summary(object, conf = 0.95, gamma = 1, l = NULL, m = 50, ...)

Arguments

object

(horseshoe)
the horseshoe object.

conf

(scalar)
the level of the credible intervals. Default is 0.95.

gamma

(scalar)
numeric value defining the weights to obtain the average hazard ratio. Default is 1 (in this case the average hazard ratio obtained can be interpreted as the odds of concordance). Just needed when using survival data.

l

(scalar)
the maximum value of time that wants to be studied to obtain the average hazard ratio. Default is the maximum value of time when there was an event. Just needed when using survival data.

m

(scalar)
the value that defines the equally spaced time points where the survival curves are going to be studied. Default is 50. Just needed when using survival data.

...

Arguments of summary.

Value

Object of class summary.horseshoe which is a list with the approximated posterior distribution of the treatment effects, a data.frame with the estimated subgroup treatment effect (with the median) and the bounds of the credible intervals, the resptype

and the confidence level.

Examples

summary(horseshoe_fit_bin)
#>   subgroup trt.estimate    trt.low   trt.high
#> 1     x_1a   -0.4633353 -0.7765820 -0.1563293
#> 2     x_1b   -0.4670046 -0.7840801 -0.1634735
#> 3     x_2a   -0.4665271 -0.7985268 -0.1476238
#> 4     x_2b   -0.4645879 -0.7674570 -0.1619634