Summary Naive

Function to obtain the naive subgroup treatment effects of an object fitted with the naive function.

Usage

# S3 method for naive
summary(object, conf = 0.95, ...)

Arguments

object

(naive)
the naive object.

conf

(scalar)
the confidence level of the intervals. Default is 0.95.

...

Arguments of summary.

Value

Object of class summary.elastic_net which is a list with a data.frame with 4 columns (the subgroup variables, the estimated treatment effect and the low and high bounds of the confidence interval of the treatment effect), the resptype and the confidence level.

Examples

summary(naive_fit_surv)
#>    subgroup trt.estimate   trt.low  trt.high
#> 1      x_1a    0.6741880 0.4693783 0.9683648
#> 2      x_1b    0.6645867 0.4663568 0.9470762
#> 3      x_2a    0.7087873 0.4743926 1.0589952
#> 4      x_2b    0.6424496 0.4636818 0.8901394
#> 5      x_3a    0.5054825 0.3003642 0.8506759
#> 6      x_3b    0.7168399 0.5365623 0.9576884
#> 7      x_4a    0.6213593 0.4133445 0.9340571
#> 8      x_4b    0.5428680 0.3372320 0.8738959
#> 9      x_4c    0.8575696 0.5483543 1.3411505
#> 10     x_5a    1.0714912 0.5843316 1.9647978
#> 11     x_5b    0.3981145 0.1911053 0.8293607
#> 12     x_5c    0.5966367 0.3719363 0.9570869
#> 13     x_5d    0.6818581 0.4569275 1.0175148
#> 14     x_6a    0.7500611 0.4833380 1.1639715
#> 15     x_6b    0.6203688 0.4551697 0.8455251
#> 16     x_7a    0.4884324 0.3137099 0.7604677
#> 17     x_7b    0.7883981 0.5775406 1.0762388
#> 18     x_8a    0.5723798 0.3253240 1.0070535
#> 19     x_8b    0.5663128 0.3505993 0.9147486
#> 20     x_8c    0.7809059 0.5485165 1.1117516
#> 21     x_9a    0.4616452 0.2454835 0.8681493
#> 22     x_9b    0.7164771 0.5426620 0.9459653
#> 23    x_10a    0.6844865 0.3818088 1.2271109
#> 24    x_10b    0.7063065 0.4515212 1.1048625
#> 25    x_10c    0.6366207 0.4435313 0.9137708